Backend/requirements.txt
fastapi
uvicorn
psycopg2
asyncpg
sqlalchemy
python-dotenv
